backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-multipart
pandas
openpyxl
google-generativeai
Pillow
httpx
